✨ added pause-resume prefix cmd

+++ commands/prefix/music/skip.js
@@ -1,7 +1,8 @@
 const { usePlayer } = require("discord-player");
 const { errorLog } = require("../../../configs/logger");
-const { requireSessionConditions, successEmbed, errorEmbed } = require("../../../helper/utils");
+const { requireSessionConditions, errorEmbed } = require("../../../helper/utils");
 const { BOT_MSGE_DELETE_TIMEOUT } = require("../../../helper/constants");
+const { success } = require("../../../configs/emojis");
 
 module.exports = {
     category: 'music',
@@ -25,33 +26,35 @@ module.exports = {
                         errorEmbed(`No music is currently being played`)
                     ]
                 }).then(msge => setTimeout(() => msge.delete(), BOT_MSGE_DELETE_TIMEOUT)).catch(err => {
-                    errorLog('An error occured with prefix skip command!')
+                    errorLog('An error occurred with prefix skip command!')
                     console.log(err);
-                })
+                });
             }
 
             const successSkip = guildPlayerNode.skip();
-            return message.reply({
-                embeds: [
-                    successSkip
-                        ? successEmbed(` Skipped **[${currentTrack}](${currentTrack.url})** song - By ${message.author}`)
-                        : errorEmbed(` Something went wrong - couldn't skip current playing song`)
-                ]
-            }).then(msge => setTimeout(() => msge.delete(), BOT_MSGE_DELETE_TIMEOUT)).catch(err => {
-                errorLog('An error occured with prefix skip command!')
-                console.log(err);
-            })
+            if (successSkip) {
+                await message.react(success); // React with success emoji
+            } else {
+                await message.reply({
+                    embeds: [
+                        errorEmbed(`Something went wrong - couldn't skip current playing song`)
+                    ]
+                }).then(msge => setTimeout(() => msge.delete(), BOT_MSGE_DELETE_TIMEOUT)).catch(err => {
+                    errorLog('An error occurred with prefix skip command!')
+                    console.log(err);
+                });
+            }
         }
         catch (error) {
-            errorLog(error.message)
+            errorLog(error.message);
             return message.reply({
                 embeds: [
                     errorEmbed(`Something went wrong while executing \`skip\` command`)
                 ],
             }).then(msge => setTimeout(() => msge.delete(), BOT_MSGE_DELETE_TIMEOUT)).catch(err => {
-                errorLog('An error occured with prefix skip command!')
+                errorLog('An error occurred with prefix skip command!')
                 console.log(err);
-            })
+            });
         }
     },
 };
